pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.7.1",
    "googletrans==4.0.0rc1",
    "gtts>=2.5.4",
    "matplotlib>=3.10.1",
    "nltk>=3.9.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
]
